refactor: API一貫性修正と肥大化ファイルの責務分離
P1: google-calendar-client.jsのrespondToEvent()で_checkResponse()を使用し、
_fetchEventsForCalendarIds()のcalendarList取得を_fetchWithAuth()に統一。
401/403がAuthenticationErrorとして正しくthrowされるようになった。

P2: calendar-management-card.js(1505行)をCalendarGroupManager(グループCRUD/
モーダル)とCalendarListRenderer(リスト描画/検索/UI状態)に分割。

P3: SidePanelUIController(1007行)からEventLoadingService、ReminderService、
ThemeService、AuthService、OnboardingServiceの5サービスを抽出。
コントローラーはオーケストレーションに専念。

P4: event-handlers.js(706行)からGoogleEventRendererとLocalEventRendererを
抽出。フェッチ/オーケストレーションとDOM生成を分離。

全て純粋なリファクタ（機能変更なし）。501テスト全通過。
