+++ src/options/components/calendar/calendar-management-card.js
@@ -1,10 +1,15 @@
 /**
  * CalendarManagementCard - The calendar management card component with group support
+ *
+ * Group CRUD and modal logic delegated to CalendarGroupManager.
+ * List rendering and UI state delegated to CalendarListRenderer.
  */
 import { CardComponent } from '../base/card-component.js';
 import { logError } from '../../../lib/utils.js';
-import { loadSelectedCalendars, saveSelectedCalendars, loadCalendarGroups, saveCalendarGroups } from '../../../lib/settings-storage.js';
+import { loadSelectedCalendars, saveSelectedCalendars, loadCalendarGroups } from '../../../lib/settings-storage.js';
 import { sendMessage } from '../../../lib/chrome-messaging.js';
+import { CalendarGroupManager } from './calendar-group-manager.js';
+import { CalendarListRenderer } from './calendar-list-renderer.js';
 
 export class CalendarManagementCard extends CardComponent {
     constructor(onCalendarSelectionChange) {
@@ -35,28 +40,37 @@ export class CalendarManagementCard extends CardComponent {
         this.clearSearchBtn = null;
         this.addGroupBtn = null;
 
-        // Active popover reference
-        this._activePopover = null;
-        this._popoverCloseHandler = null;
-        this._popoverKeyHandler = null;
-        this._popoverTimerId = null;
+        // Initialize group manager
+        this._groupManager = new CalendarGroupManager({
+            getCalendarGroups: () => this.calendarGroups,
+            setCalendarGroups: (groups) => { this.calendarGroups = groups; },
+            getSelectedCalendarIds: () => this.selectedCalendarIds,
+            setSelectedCalendarIds: (ids) => { this.selectedCalendarIds = ids; },
+            getAllCalendars: () => this.allCalendars,
+            onGroupsChanged: () => this.render(),
+            onSelectionChanged: (ids, diff) => {
+                if (this.onCalendarSelectionChange) {
+                    this.onCalendarSelectionChange(ids, diff);
+                }
+            },
+            getAddGroupBtn: () => this.addGroupBtn
+        });
 
-        // Debounced save for collapse state
-        this._collapseSaveTimer = null;
-
-        // Create-group modal references
-        this._createGroupModalOverlay = null;
-        this._createGroupModalKeyHandler = null;
-        this._isSubmittingGroup = false;
+        // Initialize list renderer
+        this._listRenderer = new CalendarListRenderer({
+            getCalendarList: () => this.calendarList,
+            getNoCalendarsMsg: () => this.noCalendarsMsg,
+            getClearSearchBtn: () => this.clearSearchBtn,
+            getSelectedCalendarIds: () => this.selectedCalendarIds,
+            getCalendarGroups: () => this.calendarGroups,
+            getAllCalendars: () => this.allCalendars,
+            getLoadingIndicator: () => this.loadingIndicator,
+            getRefreshBtn: () => this.refreshBtn
+        });
     }
 
     destroy() {
-        this._closePopover();
-        this._closeGroupModal();
-        if (this._collapseSaveTimer) {
-            clearTimeout(this._collapseSaveTimer);
-            this._collapseSaveTimer = null;
-        }
+        this._groupManager.destroy();
         super.destroy();
     }
 
@@ -221,7 +235,7 @@ export class CalendarManagementCard extends CardComponent {
         this.refreshBtn?.addEventListener('click', () => this.refreshCalendars());
 
         // The search functionality
-        this.searchInput?.addEventListener('input', (e) => this._handleSearch(e.target.value));
+        this.searchInput?.addEventListener('input', () => this.render());
         this.searchInput?.addEventListener('keyup', (e) => {
             if (e.key === 'Escape') {
                 this._clearSearch();
@@ -231,7 +245,7 @@ export class CalendarManagementCard extends CardComponent {
         this.clearSearchBtn?.addEventListener('click', () => this._clearSearch());
 
         // Add group button
-        this.addGroupBtn?.addEventListener('click', () => this._handleAddGroup());
+        this.addGroupBtn?.addEventListener('click', () => this._groupManager.handleAddGroup());
 
         // Handle the calendar checkbox using event delegation
         this.calendarList?.addEventListener('change', (e) => {
@@ -242,7 +256,7 @@ export class CalendarManagementCard extends CardComponent {
             if (groupHeader) {
                 const groupId = groupHeader.dataset.groupId;
                 if (groupId) {
-                    this._handleGroupToggle(groupId, target.checked);
+                    this._groupManager.handleGroupToggle(groupId, target.checked);
                 }
                 return;
             }
@@ -259,7 +273,7 @@ export class CalendarManagementCard extends CardComponent {
             if (collapseIcon) {
                 const groupHeader = collapseIcon.closest('.calendar-group-header');
                 if (groupHeader) {
-                    this._handleGroupCollapse(groupHeader.dataset.groupId);
+                    this._groupManager.handleGroupCollapse(groupHeader.dataset.groupId, this.calendarList);
                 }
                 return;
             }
@@ -267,7 +281,7 @@ export class CalendarManagementCard extends CardComponent {
             // Group name click (also collapse/expand, unless clicking checkbox or actions)
             const groupHeader = e.target.closest('.calendar-group-header');
             if (groupHeader && !e.target.closest('input') && !e.target.closest('.group-actions')) {
-                this._handleGroupCollapse(groupHeader.dataset.groupId);
+                this._groupManager.handleGroupCollapse(groupHeader.dataset.groupId, this.calendarList);
                 return;
             }
 
@@ -275,7 +289,7 @@ export class CalendarManagementCard extends CardComponent {
             const editBtn = e.target.closest('[data-group-edit]');
             if (editBtn) {
                 e.stopPropagation();
-                this._handleStartRenameGroup(editBtn.dataset.groupEdit);
+                this._groupManager.handleStartRenameGroup(editBtn.dataset.groupEdit);
                 return;
             }
 
@@ -283,7 +297,7 @@ export class CalendarManagementCard extends CardComponent {
             const deleteBtn = e.target.closest('[data-group-delete]');
             if (deleteBtn) {
                 e.stopPropagation();
-                this._handleDeleteGroup(deleteBtn.dataset.groupDelete);
+                this._groupManager.handleDeleteGroup(deleteBtn.dataset.groupDelete);
                 return;
             }
 
@@ -293,7 +307,7 @@ export class CalendarManagementCard extends CardComponent {
                 e.stopPropagation();
                 const calendarItem = assignBtn.closest('[data-calendar-id]');
                 if (calendarItem) {
-                    this._showGroupAssignPopover(calendarItem.dataset.calendarId, assignBtn);
+                    this._groupManager.showGroupAssignPopover(calendarItem.dataset.calendarId, assignBtn);
                 }
             }
         });
@@ -304,7 +318,7 @@ export class CalendarManagementCard extends CardComponent {
                 const groupHeader = e.target.closest('.calendar-group-header');
                 if (groupHeader && e.target === groupHeader) {
                     e.preventDefault();
-                    this._handleGroupCollapse(groupHeader.dataset.groupId);
+                    this._groupManager.handleGroupCollapse(groupHeader.dataset.groupId, this.calendarList);
                 }
             }
         });
@@ -324,7 +338,7 @@ export class CalendarManagementCard extends CardComponent {
             this.render();
         } catch (error) {
             logError('Calendar data loading', error);
-            this._showError(window.getLocalizedMessage('calendarLoadError') || 'Failed to load calendar data');
+            this._listRenderer.showError(window.getLocalizedMessage('calendarLoadError') || 'Failed to load calendar data');
         }
     }
 
@@ -352,7 +366,7 @@ export class CalendarManagementCard extends CardComponent {
      * Refresh calendar list
      */
     async refreshCalendars() {
-        this._setLoading(true);
+        this._listRenderer.setLoading(true);
 
         try {
             // Load current selections from storage to avoid overwriting with empty state
@@ -384,9 +398,9 @@ export class CalendarManagementCard extends CardComponent {
             }
         } catch (error) {
             logError('Calendar list update', error);
-            this._showError(window.getLocalizedMessage('calendarUpdateError') || `Failed to update calendars: ${error.message || 'Unknown error'}`);
+            this._listRenderer.showError(window.getLocalizedMessage('calendarUpdateError') || `Failed to update calendars: ${error.message || 'Unknown error'}`);
         } finally {
-            this._setLoading(false);
+            this._listRenderer.setLoading(false);
         }
     }
 
@@ -394,10 +408,10 @@ export class CalendarManagementCard extends CardComponent {
      * Render calendar list with group support
      */
     render() {
-        this._closePopover();
+        this._groupManager.closePopover();
 
         if (!this.allCalendars || this.allCalendars.length === 0) {
-            this._showEmptyState();
+            this._listRenderer.showEmptyState();
             return;
         }
 
@@ -409,11 +423,11 @@ export class CalendarManagementCard extends CardComponent {
             : this.allCalendars;
 
         if (filteredCalendars.length === 0 && searchTerm) {
-            this._showNoSearchResults();
+            this._listRenderer.showNoSearchResults();
             return;
         }
 
-        this._hideEmptyState();
+        this._listRenderer.hideEmptyState();
         this.calendarList.innerHTML = '';
 
         // Build a set of all calendar IDs in groups
@@ -433,7 +447,7 @@ export class CalendarManagementCard extends CardComponent {
             // Skip groups with no matching calendars during search
             if (searchTerm && groupCalendars.length === 0) continue;
 
-            const groupSection = this._createGroupSection(group, groupCalendars, searchTerm);
+            const groupSection = this._listRenderer.createGroupSection(group, groupCalendars, searchTerm);
             this.calendarList.appendChild(groupSection);
         }
 
@@ -454,882 +468,11 @@ export class CalendarManagementCard extends CardComponent {
             });
 
         if (ungroupedCalendars.length > 0) {
-            const ungroupedSection = this._createUngroupedSection(ungroupedCalendars, searchTerm);
+            const ungroupedSection = this._listRenderer.createUngroupedSection(ungroupedCalendars, searchTerm);
             this.calendarList.appendChild(ungroupedSection);
         }
 
-        this._updateSearchUI(searchTerm);
-    }
-
-    /**
-     * Create a group section (header + body)
-     * @private
-     */
-    _createGroupSection(group, calendars, searchTerm) {
-        const section = document.createElement('div');
-        section.className = 'calendar-group-section';
-
-        const header = this._createGroupHeader(group, calendars);
-        section.appendChild(header);
-
-        const body = document.createElement('div');
-        body.className = 'calendar-group-body';
-        if (group.collapsed && !searchTerm) {
-            body.classList.add('collapsed');
-        }
-
-        const sortedCalendars = [...calendars].sort((a, b) => {
-            if (a.primary && !b.primary) return -1;
-            if (!a.primary && b.primary) return 1;
-            return (a.summary || '').localeCompare(b.summary || '');
-        });
-
-        sortedCalendars.forEach(calendar => {
-            const isSelected = this.selectedCalendarIds.includes(calendar.id);
-            const item = this._createCalendarItem(calendar, isSelected);
-            body.appendChild(item);
-        });
-
-        section.appendChild(body);
-        return section;
-    }
-
-    /**
-     * Create a group header
-     * @private
-     */
-    _createGroupHeader(group, _calendarsInGroup) {
-        const header = document.createElement('div');
-        header.className = 'calendar-group-header';
-        header.dataset.groupId = group.id;
-
-        // Group checkbox
-        const checkbox = document.createElement('input');
-        checkbox.type = 'checkbox';
-        checkbox.className = 'form-check-input';
-
-        // Determine check state using full group membership (not search-filtered view)
-        const fullGroupIds = group.calendarIds.filter(id =>
-            this.allCalendars.some(cal => cal.id === id)
-        );
-        const selectedCount = fullGroupIds.filter(
-            id => this.selectedCalendarIds.includes(id)
-        ).length;
-
-        if (selectedCount === 0 || fullGroupIds.length === 0) {
-            checkbox.checked = false;
-            checkbox.setAttribute('aria-checked', 'false');
-        } else if (selectedCount === fullGroupIds.length) {
-            checkbox.checked = true;
-            checkbox.setAttribute('aria-checked', 'true');
-        } else {
-            checkbox.checked = false;
-            checkbox.indeterminate = true;
-            checkbox.setAttribute('aria-checked', 'mixed');
-        }
-
-        checkbox.setAttribute('aria-label', group.name);
-
-        // Collapse icon
-        const collapseIcon = document.createElement('i');
-        collapseIcon.className = `fas fa-chevron-down group-collapse-icon${group.collapsed ? ' collapsed' : ''}`;
-        collapseIcon.setAttribute('aria-hidden', 'true');
-
-        // Group name
-        const nameSpan = document.createElement('span');
-        nameSpan.className = 'group-name';
-        nameSpan.textContent = group.name;
-
-        // Action buttons
-        const actions = document.createElement('div');
-        actions.className = 'group-actions';
-
-        const editBtn = document.createElement('button');
-        editBtn.type = 'button';
-        editBtn.dataset.groupEdit = group.id;
-        editBtn.title = window.getLocalizedMessage('editGroup') || 'Edit';
-        editBtn.setAttribute('aria-label', editBtn.title);
-        editBtn.innerHTML = '<i class="fas fa-pencil-alt"></i>';
-
-        const deleteBtn = document.createElement('button');
-        deleteBtn.type = 'button';
-        deleteBtn.dataset.groupDelete = group.id;
-        deleteBtn.title = window.getLocalizedMessage('deleteGroup') || 'Delete';
-        deleteBtn.setAttribute('aria-label', deleteBtn.title);
-        deleteBtn.innerHTML = '<i class="fas fa-trash-alt"></i>';
-
-        actions.appendChild(editBtn);
-        actions.appendChild(deleteBtn);
-
-        header.setAttribute('role', 'button');
-        header.setAttribute('tabindex', '0');
-        header.setAttribute('aria-expanded', group.collapsed ? 'false' : 'true');
-        header.appendChild(checkbox);
-        header.appendChild(collapseIcon);
-        header.appendChild(nameSpan);
-        header.appendChild(actions);
-
-        return header;
-    }
-
-    /**
-     * Create ungrouped section
-     * @private
-     */
-    _createUngroupedSection(calendars, _searchTerm) {
-        const section = document.createElement('div');
-        section.className = 'calendar-group-section';
-
-        // Ungrouped header (simplified, no edit/delete)
-        const header = document.createElement('div');
-        header.className = 'calendar-group-header';
-        header.dataset.groupId = '__ungrouped__';
-
-        const collapseIcon = document.createElement('i');
-        collapseIcon.className = 'fas fa-chevron-down group-collapse-icon';
-        collapseIcon.setAttribute('aria-hidden', 'true');
-
-        const nameSpan = document.createElement('span');
-        nameSpan.className = 'group-name';
-        nameSpan.textContent = window.getLocalizedMessage('ungrouped') || 'Ungrouped';
-
-        // Spacer to align with grouped headers that have a checkbox
-        const spacer = document.createElement('span');
-        spacer.className = 'group-checkbox-spacer';
-        spacer.setAttribute('aria-hidden', 'true');
-
-        header.setAttribute('role', 'button');
-        header.setAttribute('tabindex', '0');
-        header.setAttribute('aria-expanded', 'true');
-        header.appendChild(spacer);
-        header.appendChild(collapseIcon);
-        header.appendChild(nameSpan);
-        section.appendChild(header);
-
-        const body = document.createElement('div');
-        body.className = 'calendar-group-body';
-
-        calendars.forEach(calendar => {
-            const isSelected = this.selectedCalendarIds.includes(calendar.id);
-            const item = this._createCalendarItem(calendar, isSelected);
-            body.appendChild(item);
-        });
-
-        section.appendChild(body);
-        return section;
-    }
-
-    /**
-     * Create calendar item
-     * @private
-     */
-    _createCalendarItem(calendar, isSelected) {
-        const item = document.createElement('div');
-        item.className = 'list-group-item d-flex align-items-center py-2';
-        item.dataset.calendarId = calendar.id;
-
-        // The checkbox
-        const checkbox = document.createElement('input');
-        checkbox.type = 'checkbox';
-        checkbox.className = 'form-check-input me-3';
-        checkbox.checked = isSelected;
-
-        checkbox.setAttribute('aria-label', calendar.summary || '');
-
-        if (calendar.primary) {
-            checkbox.disabled = true;
-            checkbox.checked = true;
-        }
-
-        // The calendar information
-        const info = document.createElement('div');
-        info.className = 'flex-grow-1';
-
-        const name = document.createElement('div');
-        name.className = 'fw-bold';
-        name.textContent = calendar.summary;
-        if (calendar.primary) {
-            name.classList.add('text-primary');
-        }
-
-        info.appendChild(name);
-
-        // Group assign button (only if groups exist)
-        const assignBtn = document.createElement('button');
-        assignBtn.type = 'button';
-        assignBtn.className = 'calendar-group-assign-btn';
-        assignBtn.title = window.getLocalizedMessage('assignToGroups') || 'Assign to groups';
-        assignBtn.setAttribute('aria-label', assignBtn.title);
-        assignBtn.innerHTML = '<i class="fas fa-folder"></i>';
-
-        // The color indicator
-        const colorIndicator = document.createElement('div');
-        colorIndicator.className = 'calendar-color-indicator me-2';
-        if (calendar.backgroundColor) {
-            colorIndicator.style.backgroundColor = calendar.backgroundColor;
-        }
-
-        item.appendChild(checkbox);
-        item.appendChild(info);
-        if (this.calendarGroups.length > 0 && !calendar.primary) {
-            item.appendChild(assignBtn);
-        }
-        item.appendChild(colorIndicator);
-
-        return item;
-    }
-
-    /**
-     * Show group assignment popover for a calendar
-     * @private
-     */
-    _showGroupAssignPopover(calendarId, anchorElement) {
-        this._closePopover();
-
-        const popover = document.createElement('div');
-        popover.className = 'calendar-group-assign-popover';
-        popover.setAttribute('role', 'dialog');
-        popover.setAttribute('aria-label', window.getLocalizedMessage('assignToGroups') || 'Assign to groups');
-
-        if (this.calendarGroups.length === 0) {
-            popover.textContent = window.getLocalizedMessage('noGroupsAvailable') || 'No groups available';
-            popover.setAttribute('tabindex', '-1');
-        } else {
-            this.calendarGroups.forEach(group => {
-                const wrapper = document.createElement('div');
-                wrapper.className = 'form-check';
-
-                const checkbox = document.createElement('input');
-                checkbox.type = 'checkbox';
-                checkbox.className = 'form-check-input';
-                checkbox.id = `assign-${calendarId}-${group.id}`;
-                checkbox.checked = group.calendarIds.includes(calendarId);
-                checkbox.addEventListener('change', () => {
-                    this._handleCalendarGroupAssignment(calendarId, group.id, checkbox.checked);
-                });
-
-                const label = document.createElement('label');
-                label.className = 'form-check-label';
-                label.htmlFor = checkbox.id;
-                label.textContent = group.name;
-
-                wrapper.appendChild(checkbox);
-                wrapper.appendChild(label);
-                popover.appendChild(wrapper);
-            });
-        }
-
-        // Position popover using fixed positioning to avoid overflow clipping
-        const rect = anchorElement.getBoundingClientRect();
-        popover.style.top = `${rect.bottom + 4}px`;
-        popover.style.right = `${document.documentElement.clientWidth - rect.right}px`;
-        document.body.appendChild(popover);
-
-        this._activePopover = popover;
-
-        // Escape key closes popover (document-level so it works even after Tab-out)
-        this._popoverKeyHandler = (e) => {
-            if (e.key === 'Escape') {
-                this._closePopover();
-                anchorElement.focus();
-            }
-        };
-        document.addEventListener('keydown', this._popoverKeyHandler);
-
-        // Focus first interactive element in popover
-        const focusTarget = popover.querySelector('input[type="checkbox"]') || popover;
-        setTimeout(() => focusTarget.focus(), 0);
-
-        this._popoverCloseHandler = (e) => {
-            if (!this._activePopover) return;
-            if (!popover.contains(e.target) && !anchorElement.contains(e.target)) {
-                this._closePopover();
-            }
-        };
-        this._popoverTimerId = setTimeout(() => {
-            this._popoverTimerId = null;
-            if (this._activePopover === popover) {
-                document.addEventListener('click', this._popoverCloseHandler);
-            }
-        }, 0);
-    }
-
-    /**
-     * Close active popover
-     * @private
-     */
-    _closePopover() {
-        if (this._popoverTimerId) {
-            clearTimeout(this._popoverTimerId);
-            this._popoverTimerId = null;
-        }
-        if (this._popoverKeyHandler) {
-            document.removeEventListener('keydown', this._popoverKeyHandler);
-            this._popoverKeyHandler = null;
-        }
-        if (this._activePopover) {
-            this._activePopover.remove();
-            this._activePopover = null;
-        }
-        if (this._popoverCloseHandler) {
-            document.removeEventListener('click', this._popoverCloseHandler);
-            this._popoverCloseHandler = null;
-        }
-    }
-
-    /**
-     * Handle calendar group assignment change
-     * @private
-     */
-    async _handleCalendarGroupAssignment(calendarId, groupId, assigned) {
-        const primaryCalendar = this.allCalendars.find(c => c.primary);
-        if (primaryCalendar && primaryCalendar.id === calendarId) return;
-        const group = this.calendarGroups.find(g => g.id === groupId);
-        if (!group) return;
-
-        const previousCalendarIds = [...group.calendarIds];
-
-        if (assigned) {
-            if (!group.calendarIds.includes(calendarId)) {
-                group.calendarIds.push(calendarId);
-            }
-        } else {
-            group.calendarIds = group.calendarIds.filter(id => id !== calendarId);
-        }
-
-        try {
-            await saveCalendarGroups(this.calendarGroups);
-            this._closePopover();
-            this.render();
-        } catch (error) {
-            group.calendarIds = previousCalendarIds;
-            this._closePopover();
-            this.render();
-            logError('Calendar group assignment save', error);
-        }
-    }
-
-    /**
-     * Handle adding a new group — opens modal
-     * @private
-     */
-    _handleAddGroup() {
-        this._showGroupModal(null);
-    }
-
-    /**
-     * Show the group modal for create or edit
-     * @param {Object|null} editingGroup - existing group to edit, or null for create
-     * @private
-     */
-    _showGroupModal(editingGroup) {
-        this._groupModalTrigger = document.activeElement;
-        this._closePopover();
-        this._closeGroupModal();
-        this._isSubmittingGroup = false;
-
-        const isEdit = !!editingGroup;
-        const modalTitle = isEdit
-            ? (window.getLocalizedMessage('editGroupTitle') || 'Edit Group')
-            : (window.getLocalizedMessage('createGroupTitle') || 'Create Group');
-        const submitLabel = isEdit
-            ? (window.getLocalizedMessage('saveGroupButton') || 'Save')
-            : (window.getLocalizedMessage('createGroupButton') || 'Create');
-        const existingCalendarIds = isEdit ? new Set(editingGroup.calendarIds) : new Set();
-
-        // Overlay
-        const overlay = document.createElement('div');
-        overlay.className = 'create-group-modal-overlay';
-
-        // Modal container
-        const modal = document.createElement('div');
-        modal.className = 'create-group-modal';
-        modal.setAttribute('role', 'dialog');
-        modal.setAttribute('aria-modal', 'true');
-        modal.setAttribute('aria-label', modalTitle);
-
-        // Header
-        const header = document.createElement('div');
-        header.className = 'create-group-modal-header';
-        const title = document.createElement('h5');
-        title.textContent = modalTitle;
-        const closeBtn = document.createElement('button');
-        closeBtn.type = 'button';
-        closeBtn.className = 'btn-close';
-        closeBtn.setAttribute('aria-label', window.getLocalizedMessage('close') || 'Close');
-        closeBtn.addEventListener('click', () => this._closeGroupModal());
-        header.appendChild(title);
-        header.appendChild(closeBtn);
-
-        // Body
-        const body = document.createElement('div');
-        body.className = 'create-group-modal-body';
-
-        // Group name input
-        const nameLabel = document.createElement('label');
-        nameLabel.className = 'form-label fw-bold';
-        nameLabel.htmlFor = 'create-group-name-input';
-        nameLabel.textContent = window.getLocalizedMessage('groupNameLabel') || 'Group Name';
-        const nameInput = document.createElement('input');
-        nameInput.type = 'text';
-        nameInput.id = 'create-group-name-input';
-        nameInput.className = 'form-control mb-3';
-        nameInput.placeholder = window.getLocalizedMessage('groupNamePlaceholder') || 'Enter group name';
-        nameInput.maxLength = 50;
-        if (isEdit) {
-            nameInput.value = editingGroup.name;
-        }
-
-        body.appendChild(nameLabel);
-        body.appendChild(nameInput);
-
-        // Calendar selection
-        const calLabel = document.createElement('label');
-        calLabel.className = 'form-label fw-bold';
-        calLabel.textContent = window.getLocalizedMessage('selectCalendarsLabel') || 'Select Calendars';
-        body.appendChild(calLabel);
-
-        // Selected calendars chip area
-        const chipArea = document.createElement('div');
-        chipArea.className = 'create-group-modal-chip-area';
-        body.appendChild(chipArea);
-
-        // Search input for calendar filtering
-        const searchInput = document.createElement('input');
-        searchInput.type = 'text';
-        searchInput.className = 'form-control form-control-sm mb-2';
-        searchInput.placeholder = window.getLocalizedMessage('searchCalendars') || 'Search calendars...';
-        searchInput.setAttribute('aria-label', window.getLocalizedMessage('searchCalendars') || 'Search calendars');
-        body.appendChild(searchInput);
-
-        const calList = document.createElement('div');
-        calList.className = 'create-group-modal-calendar-list';
-
-        // Exclude primary calendar — it is always visible and not assignable to groups
-        const sortedCalendars = this.allCalendars
-            .filter(c => !c.primary)
-            .sort((a, b) => (a.summary || '').localeCompare(b.summary || ''));
-
-        // Map calendar ID → info for chip rendering
-        const calendarInfoMap = new Map();
-        for (const cal of sortedCalendars) {
-            calendarInfoMap.set(cal.id, { name: cal.summary || cal.id, color: cal.backgroundColor || '' });
-        }
-
-        const checkboxes = [];
-        const calItems = [];
-
-        // Render chip area from current selection
-        const renderChips = () => {
-            chipArea.innerHTML = '';
-            const selectedIds = checkboxes.filter(cb => cb.checked).map(cb => cb.value);
-            if (selectedIds.length === 0) {
-                chipArea.style.display = 'none';
-                return;
-            }
-            chipArea.style.display = '';
-            for (const id of selectedIds) {
-                const info = calendarInfoMap.get(id);
-                if (!info) continue;
-                const chip = document.createElement('span');
-                chip.className = 'create-group-modal-chip';
-
-                if (info.color) {
-                    const dot = document.createElement('span');
-                    dot.className = 'calendar-color-indicator-inline';
-                    dot.style.backgroundColor = info.color;
-                    chip.appendChild(dot);
-                }
-
-                const nameText = document.createElement('span');
-                nameText.textContent = info.name;
-                chip.appendChild(nameText);
-
-                const removeBtn = document.createElement('button');
-                removeBtn.type = 'button';
-                removeBtn.className = 'create-group-modal-chip-remove';
-                removeBtn.setAttribute('aria-label', `${window.getLocalizedMessage('removeCalendar') || 'Remove'} ${info.name}`);
-                removeBtn.innerHTML = '<i class="fas fa-times"></i>';
-                removeBtn.addEventListener('click', () => {
-                    // Uncheck the corresponding checkbox
-                    const cb = checkboxes.find(c => c.value === id);
-                    if (cb) {
-                        cb.checked = false;
-                    }
-                    renderChips();
-                });
-                chip.appendChild(removeBtn);
-                chipArea.appendChild(chip);
-            }
-        };
-
-        if (sortedCalendars.length === 0) {
-            const empty = document.createElement('div');
-            empty.className = 'text-muted p-2';
-            empty.textContent = window.getLocalizedMessage('noCalendarsToAdd')
-                || 'No calendars available. Refresh the calendar list first.';
-            calList.appendChild(empty);
-        } else {
-            for (const cal of sortedCalendars) {
-                const wrapper = document.createElement('div');
-                wrapper.className = 'form-check create-group-modal-cal-item';
-
-                const checkbox = document.createElement('input');
-                checkbox.type = 'checkbox';
-                checkbox.className = 'form-check-input';
-                checkbox.id = `create-group-cal-${cal.id}`;
-                checkbox.value = cal.id;
-                if (existingCalendarIds.has(cal.id)) {
-                    checkbox.checked = true;
-                }
-
-                checkbox.addEventListener('change', () => renderChips());
-
-                const label = document.createElement('label');
-                label.className = 'form-check-label';
-                label.htmlFor = checkbox.id;
-
-                const colorDot = document.createElement('span');
-                colorDot.className = 'calendar-color-indicator-inline';
-                if (cal.backgroundColor) {
-                    colorDot.style.backgroundColor = cal.backgroundColor;
-                }
-
-                const nameSpan = document.createElement('span');
-                nameSpan.textContent = cal.summary || cal.id;
-
-                label.appendChild(colorDot);
-                label.appendChild(nameSpan);
-                wrapper.appendChild(checkbox);
-                wrapper.appendChild(label);
-                calList.appendChild(wrapper);
-                checkboxes.push(checkbox);
-                calItems.push({ element: wrapper, name: (cal.summary || cal.id).toLowerCase() });
-            }
-        }
-        body.appendChild(calList);
-
-        // Initial chip render
-        renderChips();
-
-        // Filter calendar items as user types
-        searchInput.addEventListener('input', () => {
-            const term = searchInput.value.toLowerCase().trim();
-            for (const item of calItems) {
-                item.element.style.display = (!term || item.name.includes(term)) ? '' : 'none';
-            }
-        });
-
-        // Footer
-        const footer = document.createElement('div');
-        footer.className = 'create-group-modal-footer';
-
-        const cancelBtn = document.createElement('button');
-        cancelBtn.type = 'button';
-        cancelBtn.className = 'btn btn-outline-secondary btn-sm';
-        cancelBtn.textContent = window.getLocalizedMessage('cancelButton') || 'Cancel';
-        cancelBtn.addEventListener('click', () => this._closeGroupModal());
-
-        const submitBtn = document.createElement('button');
-        submitBtn.type = 'button';
-        submitBtn.className = 'btn btn-primary btn-sm';
-        submitBtn.textContent = submitLabel;
-        submitBtn.addEventListener('click', () => {
-            this._submitGroupModal(nameInput.value.trim(), checkboxes, editingGroup);
-        });
-
-        footer.appendChild(cancelBtn);
-        footer.appendChild(submitBtn);
-
-        modal.appendChild(header);
-        modal.appendChild(body);
-        modal.appendChild(footer);
-        overlay.appendChild(modal);
-
-        // Keyboard handler: Escape to close + focus trap
-        this._createGroupModalKeyHandler = (e) => {
-            if (e.key === 'Escape') {
-                this._closeGroupModal();
-                return;
-            }
-            if (e.key === 'Tab') {
-                const focusable = Array.from(modal.querySelectorAll(
-                    'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])'
-                )).filter(el => el.offsetParent !== null);
-                if (focusable.length === 0) return;
-                const first = focusable[0];
-                const last = focusable[focusable.length - 1];
-                if (e.shiftKey) {
-                    if (document.activeElement === first) {
-                        e.preventDefault();
-                        last.focus();
-                    }
-                } else {
-                    if (document.activeElement === last) {
-                        e.preventDefault();
-                        first.focus();
-                    }
-                }
-            }
-        };
-        document.addEventListener('keydown', this._createGroupModalKeyHandler);
-
-        // Click overlay to close
-        overlay.addEventListener('mousedown', (e) => {
-            if (e.target === overlay) {
-                this._closeGroupModal();
-            }
-        });
-
-        // Enter key in name input submits
-        nameInput.addEventListener('keydown', (e) => {
-            if (e.key === 'Enter') {
-                e.preventDefault();
-                this._submitGroupModal(nameInput.value.trim(), checkboxes, editingGroup);
-            }
-        });
-
-        this._createGroupModalOverlay = overlay;
-        document.body.appendChild(overlay);
-
-        // Focus name input
-        setTimeout(() => nameInput.focus(), 0);
-    }
-
-    /**
-     * Submit the group modal (create or edit)
-     * @private
-     */
-    async _submitGroupModal(name, checkboxes, editingGroup) {
-        if (this._isSubmittingGroup) return;
-        this._isSubmittingGroup = true;
-
-        const groupName = name || (window.getLocalizedMessage('newGroupName') || 'New Group');
-        const selectedCalIds = checkboxes
-            .filter(cb => cb.checked)
-            .map(cb => cb.value);
-
-        if (editingGroup) {
-            // Edit mode: update existing group
-            const previousName = editingGroup.name;
-            const previousCalendarIds = [...editingGroup.calendarIds];
-            editingGroup.name = groupName.slice(0, 50);
-            editingGroup.calendarIds = selectedCalIds;
-
-            try {
-                await saveCalendarGroups(this.calendarGroups);
-                this._closeGroupModal();
-                this.render();
-            } catch (error) {
-                editingGroup.name = previousName;
-                editingGroup.calendarIds = previousCalendarIds;
-                logError('Edit group', error);
-            } finally {
-                this._isSubmittingGroup = false;
-            }
-        } else {
-            // Create mode: add new group
-            const groupId = `group_${Date.now()}_${Math.random().toString(36).slice(2, 8)}`;
-            const newGroup = {
-                id: groupId,
-                name: groupName.slice(0, 50),
-                calendarIds: selectedCalIds,
-                collapsed: false
-            };
-            this.calendarGroups.push(newGroup);
-
-            try {
-                await saveCalendarGroups(this.calendarGroups);
-                this._closeGroupModal();
-                this.render();
-            } catch (error) {
-                this.calendarGroups = this.calendarGroups.filter(g => g.id !== groupId);
-                logError('Add group', error);
-            } finally {
-                this._isSubmittingGroup = false;
-            }
-        }
-    }
-
-    /**
-     * Close the group modal
-     * @private
-     */
-    _closeGroupModal() {
-        const wasOpen = !!this._createGroupModalOverlay;
-        if (this._createGroupModalKeyHandler) {
-            document.removeEventListener('keydown', this._createGroupModalKeyHandler);
-            this._createGroupModalKeyHandler = null;
-        }
-        if (this._createGroupModalOverlay) {
-            this._createGroupModalOverlay.remove();
-            this._createGroupModalOverlay = null;
-        }
-        if (wasOpen) {
-            const trigger = this._groupModalTrigger;
-            this._groupModalTrigger = null;
-            if (trigger && trigger.isConnected) {
-                trigger.focus();
-            } else if (this.addGroupBtn && this.addGroupBtn.isConnected) {
-                this.addGroupBtn.focus();
-            }
-        }
-    }
-
-    /**
-     * Handle deleting a group
-     * @private
-     */
-    async _handleDeleteGroup(groupId) {
-        const confirmMsg = window.getLocalizedMessage('deleteGroupConfirm')
-            || 'Delete this group? Calendars will be moved to Ungrouped.';
-        if (!confirm(confirmMsg)) return;
-
-        const previousGroups = [...this.calendarGroups];
-        this.calendarGroups = this.calendarGroups.filter(g => g.id !== groupId);
-
-        try {
-            await saveCalendarGroups(this.calendarGroups);
-            this.render();
-        } catch (error) {
-            // Rollback on save failure
-            this.calendarGroups = previousGroups;
-            logError('Delete group', error);
-        }
-    }
-
-    /**
-     * Open edit modal for a group
-     * @private
-     */
-    _handleStartRenameGroup(groupId) {
-        const group = this.calendarGroups.find(g => g.id === groupId);
-        if (!group) return;
-        this._showGroupModal(group);
-    }
-
-    /**
-     * Handle group toggle (select/deselect all calendars in group)
-     * @private
-     */
-    async _handleGroupToggle(groupId, isChecked) {
-        const group = this.calendarGroups.find(g => g.id === groupId);
-        if (!group || group.calendarIds.length === 0) return;
-
-        const previousIds = [...this.selectedCalendarIds];
-        const primaryId = this.allCalendars.find(c => c.primary)?.id;
-
-        if (isChecked) {
-            // Add all valid group calendar IDs not already selected
-            const allCalendarIds = new Set(this.allCalendars.map(c => c.id));
-            for (const calId of group.calendarIds) {
-                if (allCalendarIds.has(calId) && !this.selectedCalendarIds.includes(calId)) {
-                    this.selectedCalendarIds.push(calId);
-                }
-            }
-        } else {
-            // Remove group calendar IDs, but keep those also in other selected groups, and keep primary
-            const otherGroupSelectedIds = this._getOtherGroupSelectedIds(groupId);
-            this.selectedCalendarIds = this.selectedCalendarIds.filter(id => {
-                if (id === primaryId) return true;
-                if (!group.calendarIds.includes(id)) return true;
-                if (otherGroupSelectedIds.has(id)) return true;
-                return false;
-            });
-        }
-
-        try {
-            await saveSelectedCalendars(this.selectedCalendarIds);
-            if (this.onCalendarSelectionChange) {
-                const addedIds = this.selectedCalendarIds.filter(id => !previousIds.includes(id));
-                const removedIds = previousIds.filter(id => !this.selectedCalendarIds.includes(id));
-                this.onCalendarSelectionChange(this.selectedCalendarIds, { addedIds, removedIds });
-            }
-            this.render();
-        } catch (error) {
-            this.selectedCalendarIds = previousIds;
-            this.render();
-            logError('Group toggle save', error);
-        }
-    }
-
-    /**
-     * Get calendar IDs that are selected and belong to other active groups
-     * @private
-     */
-    _getOtherGroupSelectedIds(excludeGroupId) {
-        const ids = new Set();
-        for (const group of this.calendarGroups) {
-            if (group.id === excludeGroupId) continue;
-            for (const calId of group.calendarIds) {
-                if (this.selectedCalendarIds.includes(calId)) {
-                    ids.add(calId);
-                }
-            }
-        }
-        return ids;
-    }
-
-    /**
-     * Handle group collapse/expand
-     * @private
-     */
-    async _handleGroupCollapse(groupId) {
-        if (groupId === '__ungrouped__') {
-            // Toggle ungrouped section (UI only, no persistence)
-            const header = this.calendarList.querySelector('.calendar-group-header[data-group-id="__ungrouped__"]');
-            if (header) {
-                const body = header.nextElementSibling;
-                const icon = header.querySelector('.group-collapse-icon');
-                if (body) body.classList.toggle('collapsed');
-                if (icon) icon.classList.toggle('collapsed');
-                const expanded = !body?.classList.contains('collapsed');
-                header.setAttribute('aria-expanded', String(expanded));
-            }
-            return;
-        }
-
-        const group = this.calendarGroups.find(g => g.id === groupId);
-        if (!group) return;
-
-        group.collapsed = !group.collapsed;
-
-        // Update UI directly without full re-render
-        const header = this.calendarList.querySelector(`.calendar-group-header[data-group-id="${CSS.escape(groupId)}"]`);
-        if (header) {
-            const body = header.nextElementSibling;
-            const icon = header.querySelector('.group-collapse-icon');
-            if (body) body.classList.toggle('collapsed', group.collapsed);
-            if (icon) icon.classList.toggle('collapsed', group.collapsed);
-            header.setAttribute('aria-expanded', group.collapsed ? 'false' : 'true');
-        }
-
-        this._debouncedSaveGroups();
-    }
-
-    /**
-     * Debounced save for calendar groups (used for collapse state)
-     * @private
-     */
-    _debouncedSaveGroups() {
-        if (this._collapseSaveTimer) {
-            clearTimeout(this._collapseSaveTimer);
-        }
-        this._collapseSaveTimer = setTimeout(async () => {
-            try {
-                await saveCalendarGroups(this.calendarGroups);
-            } catch (error) {
-                logError('Group save (debounced)', error);
-            }
-        }, 1000);
-    }
-
-    /**
-     * Handle search
-     * @private
-     */
-    _handleSearch() {
-        this.render();
+        this._listRenderer.updateSearchUI(searchTerm);
     }
 
     /**
@@ -1377,7 +520,7 @@ export class CalendarManagementCard extends CardComponent {
             this.selectedCalendarIds = previousIds;
             this.render();
             logError('Calendar selection save', error);
-            this._showError(window.getLocalizedMessage('calendarSaveError') || 'Failed to save settings');
+            this._listRenderer.showError(window.getLocalizedMessage('calendarSaveError') || 'Failed to save settings');
         }
     }
 
@@ -1413,85 +556,6 @@ export class CalendarManagementCard extends CardComponent {
                 checkbox.indeterminate = true;
                 checkbox.setAttribute('aria-checked', 'mixed');
             }
-        }
-    }
-
-    /**
-     * Set loading state
-     * @private
-     */
-    _setLoading(loading) {
-        if (this.loadingIndicator) {
-            this.loadingIndicator.style.display = loading ? 'block' : 'none';
-        }
-        if (this.refreshBtn) {
-            this.refreshBtn.disabled = loading;
-        }
-    }
-
-    /**
-     * Show empty state
-     * @private
-     */
-    _showEmptyState() {
-        this.calendarList.innerHTML = '';
-        this.noCalendarsMsg.style.display = 'block';
-    }
-
-    /**
-     * Show no search results
-     * @private
-     */
-    _showNoSearchResults() {
-        const noResultsMsg = window.getLocalizedMessage('noSearchResults') || 'No search results found';
-        this.calendarList.innerHTML = '';
-        const div = document.createElement('div');
-        div.className = 'text-muted text-center p-3';
-        div.textContent = noResultsMsg;
-        this.calendarList.appendChild(div);
-        this.noCalendarsMsg.style.display = 'none';
-    }
-
-    /**
-     * Hide empty state
-     * @private
-     */
-    _hideEmptyState() {
-        this.noCalendarsMsg.style.display = 'none';
-    }
-
-    /**
-     * Update search UI
-     * @private
-     */
-    _updateSearchUI(searchTerm) {
-        if (this.clearSearchBtn) {
-            this.clearSearchBtn.style.display = searchTerm ? 'block' : 'none';
-        }
-    }
-
-    /**
-     * Show error
-     * @private
-     */
-    _showError(message) {
-        const errorDiv = document.createElement('div');
-        errorDiv.className = 'alert alert-danger alert-dismissible fade show';
-
-        const strong = document.createElement('strong');
-        strong.textContent = (window.getLocalizedMessage('errorLabel') || 'Error') + ': ';
-        errorDiv.appendChild(strong);
-        errorDiv.appendChild(document.createTextNode(message));
-
-        const closeBtn = document.createElement('button');
-        closeBtn.type = 'button';
-        closeBtn.className = 'btn-close';
-        closeBtn.setAttribute('data-bs-dismiss', 'alert');
-        closeBtn.setAttribute('aria-label', window.getLocalizedMessage('close') || 'Close');
-        errorDiv.appendChild(closeBtn);
-
-        if (this.calendarList?.parentElement) {
-            this.calendarList.parentElement.insertBefore(errorDiv, this.calendarList);
         }
     }
